LevelBlockoutAdditions
Added some models in (Lantern, TV, Chair, Toolbox, Car (untextured), Trees, Toilet (Untextured), Sink, and Hammer. Started to add in wall textures. Added in zombie to a folder but needs to be set up. Not done adding textures to the walls. Planning on adding some rain to the outside of the house.